+++ .circleci/config.yml
@@ -2,7 +2,7 @@ version: 2
 jobs:
     build:
         docker:
-            - image: circleci/node:8.12.0
+            - image: circleci/node:10.13.0
         steps:
             - checkout
             - restore_cache:
